Track and optimize based on possible old tape values.
Some states will only encounter a '0' on the tape, or they will only encounter a '1' on the tape. If there are two such states, one for '0' and one for '1', they can be combined since the value on the tape serves to distinguish the two states.

In this change only jump states are taking advantage of the old tape values. Almost all other states expect to encounter both '0' and '1', but jump states are only entered from a limited set of PCs.

To record the old_tape values the jump builder needs to know the source PC. This initially results in additional states being built since each jump is now a distinct subroutine. But compress() can easily combine the states (updating old_tape as it goes) after which combine_states() still has the information it needs.

The optimization makes a complete mess of the jump state machine: in effect the values found in PC are feeding into the state machine to determine which jump to perform. But the result is very efficient jumps.

ZFC:       565
Loader:    875
BLC BMS:   585
Direct BMS:504